Some changes for recipe page!
Some minor changes to recipe page, can take a number at the end as
identifier for ID (0 is the example one I made so far).

Some issues with recipe page:
- Cannot grab the name of the author yet!
- Description and Ingredients not yet utilized and shown effectively
- Cannot order
- Some issues with interface/graphics, must touch up!
- It always shows that you are logged in, even if you are not!

Overall things left to do (for me):
- new recipe page is hooked up to router but cannot add any new recipe
yet! On that note, must find out how to keep track of which recipe id I
am on, or to auto increment in mongoose. Also, must be able to store
the current user’s id
- Order page and my orders page not yet hooked up! Shouldn’t be
difficult though and quite straight forward.